fix compiler warning
git-svn-id: svn+ssh://software.sandia.gov/svn/private/PowerAPI/src/prototype@597 bcb1eb47-8062-464b-8553-523ad38f42de

+++ main.c
@@ -110,13 +110,15 @@ int main( int argc, char* argv[] )
 
 
     printf("PWR_StatGetValue(PWR_ATTR_POWER) value=%f\n", value );
-    printf("PWR_StatGetValue(PWR_ATTR_POWER) start=%llu\n", statTimes.start );
+    printf("PWR_StatGetValue(PWR_ATTR_POWER) start=%llu\n", 
+                                    (long long)statTimes.start );
 
-   	printf("PWR_StatGetValue(PWR_ATTR_POWER) stop=%llu\n", statTimes.stop );
+   	printf("PWR_StatGetValue(PWR_ATTR_POWER) stop=%llu\n",
+                                    (long long) statTimes.stop );
 
 	if ( statTimes.instant != PWR_TIME_NOT_SET ) {
     	printf("PWR_StatGetValue(PWR_ATTR_POWER) instant=%llu\n",
-													statTimes.instant );
+									(long long )statTimes.instant );
 	}
 
 	PWR_StatDestroy( coreStat );
